fix order of precedence when adding tokens to the tree
There was an issue with the previous code that was not detected before
due to the number of levels of precedence between operations. Since there
were only two, namely Multiplication/Division and Addition/ Subtraction,
this problem did not arise.

Now that exponentiation is an accepted operator that is another level of
precedence, the issue arose.

The issue was in expressions that had all three levels of precedence in
sequence. For example:
> 7 + 6 * 5 ^ 2

When the algorithm reach the exponentiation expression the tree would be
this:
>  +
> / \
>7   *
>   / \
>  6   5

Since the algorithm, prior to this commit, was not recursive, it only
looked at the root of the tree, inserting the exponentiation expression
after the addition, as such:
>  +
> / \
>7   ^   <-----
>   / \
>  *   2
> / \
>6   5

This way, the evaluation algorithm would evaluate the multiplication
prior to the exponentiation.

This commit makes the algorithm that pushes tokens to the tree
recursive. Thus, whenever a new operation token has precedence over the
previous operation token, the algorithm is called passing the root's
right node as the root for the next recursion.